+++ CHANGELOG.md
@@ -42,6 +42,8 @@ MarkLogic Search Options XML Integration:
 
 1. Support resolving `parse/@at`, `start-facet/@at`, and `finish-facet/@at` attribute values in XQuery direct
    constructors to the XQuery/MJS/SJS file they reference.
+2. Support resolving `parse/@apply`, `start-facet/@apply`, and `finish-facet/@apply` attribute values in XQuery
+   direct constructors to the XQuery function they reference.
 
 ## 2021